fixed future warning by specifyiung explicitely type of empty list for pdf

+++ pySIMDEUM/core/user.py
@@ -94,7 +94,7 @@ class Presence(HasStrictTraits):
 
     def pdf(self, peak=0.65, normal=0.335, away=0.0, night=0.015):
         index = pd.timedelta_range(start='00:00:00', end='24:00:00', freq='1Min')
-        pdf = pd.Series(index=index)
+        pdf = pd.Series(index=index, dtype='float64')
 
         up = int((self.up.total_seconds()) / 60) % 1440
         up_p30 = int((up + 30)) % 1440
